SWI-Prolog 10 compatibility
Adjust FLI call sites for SWI 10's bool-returning APIs, update foreign registration pl_function_t handling, and route integer dict key lookups through get_dict/3 to avoid invalid atom_t errors.

+++ swipl/src/atom.rs
@@ -148,7 +148,7 @@ unifiable! {
     (self:Atom, term) => {
         let result = unsafe { PL_unify_atom(term.term_ptr(), self.atom) };
 
-        result != 0
+        (result as i32) != 0
     }
 }
 
@@ -169,7 +169,7 @@ where
         return Err(PrologError::Exception);
     }
 
-    let arg = if result == 0 {
+    let arg = if (result as i32) == 0 {
         None
     } else {
         let atom = unsafe { Atom::wrap(atom) };
@@ -352,7 +352,7 @@ unifiable! {
             )
         };
 
-        result != 0
+        (result as i32) != 0
     }
 }
 
@@ -383,7 +383,7 @@ where
         return Err(PrologError::Exception);
     }
 
-    let arg = if result == 0 {
+    let arg = if (result as i32) == 0 {
         None
     } else {
         let swipl_string_ref = unsafe { std::slice::from_raw_parts(ptr as *const u8, len) };
